requirements.txt
Pillow==10.1.0
PySimpleGUI==4.60.5

